// test/fixtures/conformance-schema.d.ts
export interface paths {
  '/echo/{id}': {
    parameters: {
      query?: never;
      header?: never;
      path: {
        id: string;
      };
      cookie?: never;
    };
    get: {
      parameters: {
        query?: never;
        header?: never;
        path: {
          id: string;
        };
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description Observed HTTP request */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            'application/json': {
              url: string;
              body: string;
              contentType: string;
            };
          };
        };
      };
    };
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/echo/{id}/': {
    parameters: {
      query?: never;
      header?: never;
      path: {
        id: string;
      };
      cookie?: never;
    };
    get: {
      parameters: {
        query?: never;
        header?: never;
        path: {
          id: string;
        };
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description Observed HTTP request */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            'application/json': {
              url: string;
              body: string;
              contentType: string;
            };
          };
        };
      };
    };
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/search': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get: {
      parameters: {
        query: {
          color: string[];
        };
        header?: never;
        path?: never;
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description Observed HTTP request */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            'application/json': {
              url: string;
              body: string;
              contentType: string;
            };
          };
        };
      };
    };
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/upload': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    post: {
      parameters: {
        query?: never;
        header?: never;
        path?: never;
        cookie?: never;
      };
      requestBody: {
        content: {
          'multipart/form-data': {
            value: string & unknown;
          };
        };
      };
      responses: {
        /** @description Observed HTTP request */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            'application/json': {
              url: string;
              body: string;
              contentType: string;
            };
          };
        };
      };
    };
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/reports/{year}-{month}': {
    parameters: {
      query?: never;
      header?: never;
      path: {
        year: number;
        month: number;
      };
      cookie?: never;
    };
    get: {
      parameters: {
        query?: never;
        header?: never;
        path: {
          year: number;
          month: number;
        };
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description Observed HTTP request */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            'application/json': {
              url: string;
              body: string;
              contentType: string;
            };
          };
        };
      };
    };
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
}